only install command line tools and nix if not yet installed

+++ install.sh
@@ -1,23 +1,27 @@
 #! /bin/sh
 
-# inform softwareupdate to fetch command line tools
-touch /tmp/.com.apple.dt.CommandLineTools.installondemand.in-progress
+if [ -z "$(xcode-select --print-path 2> /dev/null)" ];
+then
+  # inform softwareupdate to fetch command line tools
+  touch /tmp/.com.apple.dt.CommandLineTools.installondemand.in-progress
 
-# list software updates and return the latest command line tools (skipping potential beta listed first)
-PRODUCT=$(softwareupdate --list | grep "^\*.* Command Line Tools" | tail -n 1 | sed 's/^[^C]* //')
-echo "$PRODUCT"
+  # list software updates and return the latest command line tools (skipping potential beta listed first)
+  PRODUCT=$(softwareupdate --list | grep "^\*.* Command Line Tools" | tail -n 1 | sed 's/^[^C]* //')
 
-# download and install command line tools
-softwareupdate --install --no-scan "$PRODUCT"
+  # download and install command line tools
+  softwareupdate --install --no-scan "$PRODUCT"
 
-# remove temporary file
-rm /tmp/.com.apple.dt.CommandLineTools.installondemand.in-progress
+  # remove temporary file
+  rm /tmp/.com.apple.dt.CommandLineTools.installondemand.in-progress
+fi
 
-# download and install nix
-curl -fsSL https://nixos.org/nix/install | sh -s -- --yes
+if ! type "nix" > /dev/null; then
+  # download and install nix
+  curl -fsSL https://nixos.org/nix/install | sh -s -- --yes
 
-# load system-wide profile changes from nix
-source /etc/zprofile && source /etc/zshrc
+  # load system-wide profile changes from nix
+  . /etc/zprofile && . /etc/zshrc
+fi
 
 # clone the nix config repository
 nix-shell -p git --run "git clone https://github.com/epiccoolguy/autonix $HOME/.config/nix"
